requirements.txt
dashscope
faiss-cpu
gradio==3.47.1
langchain
markdown-cjk-spacing
mdx_truly_sane_lists
pymdown-extensions
python-slugify
unstructured

dashscope
datasets>=2.8.0
ipython
jupyter>=1.0.0
langchain<=0.0.292
modelscope>=1.7.0
moviepy
ms-swift
openai
opencv-python
openpyxl
Pillow
pydantic>=2.0.0
pypdf
pytest
python-dotenv
seaborn
soundfile
transformers>=4.29.0
transformers_stream_generator


readability-lxml
